Fixed crash when trying to delete non existing customers

+++ H1-ERP/DataBase/DB-Customers.cs
@@ -1,9 +1,5 @@
 ﻿using H1_ERP.DomainModel;
-using System.Data.Entity.Core.Metadata.Edm;
 using System.Data.SqlClient;
-using System.Data.SqlTypes;
-using System.Reflection.Metadata;
-using System.Runtime.Loader;
 
 namespace H1_ERP.DataBase
 {
@@ -13,9 +9,9 @@ namespace H1_ERP.DataBase
         {
             //for a customer obejct we're gonna need an adress aswell for the customer 
             Customer result = new Customer();
-            result.CustomerId = id; 
+            result.CustomerId = id;
             Address address = new Address();
-            SqlConnection connection = getConnection(); 
+            SqlConnection connection = getConnection();
             connection.Open();
             //Get the details of the Customer from the person table
             string sql = $"SELECT * FROM [H1PD021123_Gruppe4].[dbo].[Customers.Person] WHERE PersonID = {id}";
@@ -37,7 +33,7 @@ namespace H1_ERP.DataBase
             reader = command.ExecuteReader();
             while (reader.Read())
             {
-                address.AdressID =Convert.ToUInt32(reader.GetInt32(0));
+                address.AdressID = Convert.ToUInt32(reader.GetInt32(0));
                 address.RoadName = reader.GetString(1);
                 address.StreetNumber = reader.GetString(2);
                 address.ZipCode = reader.GetString(3);
@@ -82,9 +78,9 @@ namespace H1_ERP.DataBase
         public void InsertCustomer(Customer input)
         {
             SqlConnection Connection = getConnection();
-            uint adressid = InsertAddress(input.Address); 
+            uint adressid = InsertAddress(input.Address);
             Connection.Open();
-            
+
             uint personID = InsertPerson(input, adressid);
             SqlCommand sqlCommand = new SqlCommand($"INSERT INTO [H1PD021123_Gruppe4].[dbo].[Customer.Customers] (LastPurchaseDate,PersonID) VALUES(@Date,{personID})", Connection);
             sqlCommand.Parameters.AddWithValue("@Date", input.LastPurchaseDate);
@@ -105,20 +101,22 @@ namespace H1_ERP.DataBase
         }
 
 
-        public void DeleteCustomer(int  input)
+        public void DeleteCustomer(int input)
         {
             SqlConnection connection = getConnection();
             connection.Open();
             Customer customercustomer = GetCustomerFromID(input);
             string sqlgetrfkid = $"SELECT OrderID FROM [dbo].[Sales.Orders] WHERE CustomerID = {customercustomer.CustomerId}";
             SqlCommand command = new SqlCommand(sqlgetrfkid, connection);
-            int OrderID = 0; 
+            int OrderID = 0;
             SqlDataReader reader = command.ExecuteReader();
-            while(reader.Read())
+            while (reader.Read())
             {
                 OrderID = Convert.ToInt32(reader[0]);
             }
-            reader.Close(); 
+            reader.Close();
+
+            if (OrderID == 0) return;
 
             command = new SqlCommand($"DELETE FROM [dbo].[Sales.Orderlines] WHERE OrderID = {OrderID}", connection);
             command.ExecuteNonQuery();
@@ -128,7 +126,7 @@ namespace H1_ERP.DataBase
             command.ExecuteNonQuery();
             command = new SqlCommand($"DELETE FROM [H1PD021123_Gruppe4].[dbo].[Customers.Person] WHERE PersonID = {customercustomer.PersonID}", connection);
             command.ExecuteNonQuery();
-            command = new SqlCommand($"DELETE FROM [H1PD021123_Gruppe4].[dbo].[Customer.Adress] WHERE AdressID = {customercustomer.Address.AdressID}",connection);
+            command = new SqlCommand($"DELETE FROM [H1PD021123_Gruppe4].[dbo].[Customer.Adress] WHERE AdressID = {customercustomer.Address.AdressID}", connection);
             command.ExecuteNonQuery();
             connection.Close();
         }
@@ -137,10 +135,10 @@ namespace H1_ERP.DataBase
         {
             uint AddressID = 0;
             SqlConnection connection = getConnection();
-            connection.Open(); 
-            SqlCommand command = new SqlCommand($"INSERT INTO [H1PD021123_Gruppe4].[dbo].[Customer.Adress] (RoadName, StreetNumber, ZipCode, City, Country) VALUES  ('{address.RoadName}','{address.StreetNumber}','{address.ZipCode}','{address.City}','{address.Country}')",connection);
+            connection.Open();
+            SqlCommand command = new SqlCommand($"INSERT INTO [H1PD021123_Gruppe4].[dbo].[Customer.Adress] (RoadName, StreetNumber, ZipCode, City, Country) VALUES  ('{address.RoadName}','{address.StreetNumber}','{address.ZipCode}','{address.City}','{address.Country}')", connection);
             command.ExecuteNonQuery();
-            command = new SqlCommand("SELECT TOP (1) [AdressID] FROM [H1PD021123_Gruppe4].[dbo].[Customer.Adress] ORDER BY [AdressID] desc;",connection);
+            command = new SqlCommand("SELECT TOP (1) [AdressID] FROM [H1PD021123_Gruppe4].[dbo].[Customer.Adress] ORDER BY [AdressID] desc;", connection);
             SqlDataReader reader = command.ExecuteReader();
 
             while (reader.Read())
@@ -149,7 +147,7 @@ namespace H1_ERP.DataBase
             }
             return AddressID;
         }
-        private uint InsertPerson(Customer input,uint adressid)
+        private uint InsertPerson(Customer input, uint adressid)
         {
 
             uint PersonID = 0;
@@ -168,6 +166,6 @@ namespace H1_ERP.DataBase
 
         }
 
-      
+
     }
 }
